pubpkg/requirements.txt
asyncio-nats-streaming
flask